UI work!
Added an adapter for all three recycler views in the chatpage
- console recycler
- chat recycler
- prompt recycler

The prompt recycler should only be visible when it has content (ie: prompts) in it. otherwise the chat recycler should take up the whole screen.

Added a style xml file that adds a white border to views it's applied to.