Database, Repository, and Startup logic
Built the database as an in memory database

added a function in the chatpageviewmodel that initializes this in memory database with some basic data for testing purposes, it is triggered by an options menu item in the chatpage.

Added a repository with a function for inserting data to the user save table and a livedata variable for reading the data in said table.

Added navigation logic to the chatpage fragment that navigates to the power on fragment if the usersave data meets the conditions.

Added the logic for navigating to the chatpage and updating the user save table to reflect the sequence has started.