Repository
Finished the repository getSequence method. It now returns a fully fleshed out Sequence object from the current database values.

Wrote a utility function that takes a list of PromptLines exactly like the database returns and poops out list of Set objects to be used in the creation of said Sequence object.

Wrote a unit test to test the above utility function.