Real functionality
Added methods to the view model that allow it to "display" console/chat lines as well as prompt sets.

Added a touch handler to the chat recycler view that calls a "userTouch" method in the view model when triggered. This userTouch method then displays the next line or prompt set as needed.