Prompt handling!
Added a promptSelected method to the viewModel that displays the selected prompt in the chat recycler and then displays the next script line or prompt set in the sequence.

Added the chat viewModel as a dependency for the chat adapter. This was done to enable the adapter to call the promptSelected method directly when a prompt button is pressed by the user.