Saving and restoring from a save!
Added code to the viewmodel to save progress to the UserSave database table as script lines are displayed and prompts are displayed/selected.

Created a function called restoreSave() that will populate the console and chat adapters with all the previous lines up to the currently saved position. It will also stop on a prompt if need be.

Currently in the process of updating the UserSave table to record all of the users dialogue choices.