Notifications in progress
When the user finishes a sequence the viewmodel tells the fragment to schedule an alarm for some point in the future (yet to be determined). When the alarm goes off a broadcast receiver is triggered which immediately schedules work with the work manager. Said work then updates the user save isPowered value to true and fires a notification. (the notification part doesn't work atm but the database operation does).