Animation Object Permanence
Added code to ChatRecyclerAdapter and  ConsoleRecyclerAdapter that automatically skips the CustomTextView animations if the item is not the latest one from the dataset.

In a similar vein I also added code that checks if the last item has been animated before and if so skipping.
This will result in essentially "one shot" animations so they're not constantly reanimating every time the user scrolls through the RecyclerViews.