Theme Bugfix and getUserSave() overhaul
Fixed the "activity needs to user appcompat theme bug"

Added a new suspend method called getUserSave() to the repository and renamed the old one to getUserSaveBlocking(). I also changed the viewmodel to use the new suspend function to get the UserSave from the repo instead of using the weird livedata method. This allowed me to remove the 1 second delay from the viewmodel init.

Note: the fragment still uses the livedata version of the UserSave because it can't be asynchonous.