Trigger Resource Ids
Changed the resourceId property of triggers to be a string. And added a function that takes said string and returns whatever the resource id of that resource is. This is so triggers can reference individual animations/sound effects no matter what their unique resource id happens to be on that device. Because they can and will change.