demo mode
Added infrastructure for the new demo mode feature. This will be tied to a Boolean in the UserSave called demoMode. When this boolean is set to true through the settings menu (not implemented yet) the notification that's usually scheduled when a sequence is finished is not scheduled at all. Instead a new button is made visible in the no power fragment called skip time. When skip time is pressed the notification is fired immediately thus progressing the user to the next sequence without having to wait.