demo mode and start sequence bug fix
Added the demo mode switch to the settings menu. It controls the demoMode boolean in the user save database table. Which enables a skip time button in the no power fragment to skip the wait between sequences.

Added a boolean argument to the chat page fragment that is always false except when it has been navigated to from the power on fragment. In which case it will tell the view model to display the first ScriptLine thus starting the sequence. This used to happen in restore save if the userSave.line == 0 which caused the power on fragment to only be displayed once. If the user didn't actually press the power on button then restarted the app, it would bypass the power on fragment as the sequence was essentially already started by the view model before.