Bugfix: Sequence not clearing when skip time is used
Added a private method to the view model that essentially clears it and preps it for running the next sequence. This new method is called when a sequence ends and only ever matters if the user progresses from one sequence to the next without closing the app.
And the most common cause of that is using demo mode. (Regardless the issue should be resolved now anyway)