Dealing with demo mode side effects
Added logic to the NoPowerFragment that allows it to handle the demo mode being toggled off and on in between sequences.

The NoPowerFragment now checks if an existing notification is already scheduled and cancels it when demo mode is turned off.
On the other hand if demo mode is turned on it will schedule a new notification if one doesn't already exist.