New Events!
Migrated stopSound/Anim, startSound/anim, and startSoundOneAndDone/Anim all to use events instead of observing live data. Also moved the logic for resuming looping sound/animations to the view model. As well as added two cache variables for both the last sound and last animation trigger fired. This is used by the view model for the aforementioned logic as well as being able to do one and done triggers.