Converting from Livedata to Flows
Replaced the user save live data in the repository with a flow.

Changed the StartupViewModel to convert this flow to a StateFlow with stateIn() then collect it and make it available as a StateFlow to the NoPowerFragment. This solves the issues with it getting old data when it starts up.

Also gave the same exact treatment to the ChatPageViewModel. The SettingsFragment is the one who ultimately collects from said view model.