PowerOnFragment to ChatPageFragment
Fixed the bug where the PowerOnFragments argument that it passed to the ChatPage would never return to false and thus give a false positive forever after being triggered once.

Changed the argument in question to an integer instead of a boolean, PowerOnFragment now passes the number of the current sequence when it navigates so the ChatPageViewModel can differentiate between old and new values.