Fixed restoreSave crash
restoreSave was iterating through every ScriptLine in a sequence and running it's logic on every one. This was causing a problem because it would run it's logic and subsequently display ScriptLine's that the user never actually went through. IE: with branching dialogue options. It would go through both branches of the dialogue tree. But at the same time the PromptResult table was actually correct IE: it doesn't have prompt results for those prompt sets because the user never actually went through them.

So I rewrote the restoreSave() function to be a recursive function that more or less "simulates" playing through the sequence up to the stopping point. (more detail in the code comments).