Ending Fragment + Overhauled restoreSave()
Implemented the ending credits fragment which is navigated to when the user completes the final sequence. It also resets the UserSave so they can play through again if they want to.

Overhauled restoreSave() to take into account prompts that lead to other prompts. The old version assumed it was getting a ScriptLine every time which was a problem because some prompts "nextType" was another prompt not a ScriptLine and this would break the function.