MediaPlayer permanency
The MediaPlayer for sound effects now plays through navigation with the exception of NoPower and PowerOn fragments. As well as pausing and resuming when the user leaves the app and comes back. It still fully restarts when the app is fully closed and reopened.

This was achieved by making the mediaPlayer available through the view model and releasing/pausing/resuming it from the MainActivities's lifecycles.